# Add `optimize()` benchmark