--- a/openvm/Cargo.toml
+++ b/openvm/Cargo.toml
@@ -75,6 +75,11 @@ openvm-pairing-circuit.workspace = true
 openvm-pairing-transpiler.workspace = true
 serde_cbor = "0.11.2"
 expect-test = "1.5.1"
+criterion = { version = "0.4", features = ["html_reports"] }
 
 [lib]
 bench = false # See https://github.com/bheisler/criterion.rs/issues/458
+
+[[bench]]
+name = "optimizer_benchmark"
+harness = false