```
$ cd openvm
$ cargo bench --bench optimizer_benchmark
Benchmarking optimize-keccak/optimize: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 44.2s.
optimize-keccak/optimize
                        time:   [4.3703 s 4.3856 s 4.4020 s]
                        change: [-1.2781% -0.5240% +0.2094%] (p = 0.21 > 0.05)
                        No change in performance detected.
```